+++ Source/Pawnmorphs/Esoteria/Pawnmorph.csproj
@@ -775,7 +775,6 @@
     <Compile Include="User Interface\MutationData.cs" />
     <Compile Include="Utilities\CompCacher.cs" />
     <Compile Include="Utilities\Filter.cs" />
-    <Compile Include="Utilities\GaussianRandom.cs" />
     <Compile Include="Utilities\HediffCompBase.cs" />
     <Compile Include="Utilities\HediffDefUtilities.cs" />
     <Compile Include="Utilities\LinqUtils.cs" />
